Frontend project structure to improve readability
