+++ react-native/app/tabs/homescreen.tsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { View, StyleSheet, Text, ActivityIndicator, Dimensions } from 'react-native';
 import { Icon } from '@rneui/themed';
 import { CopilotStep } from "react-native-copilot";
-import navigate from '@/services/navigateButton';
+import navigate from '@/services/navigate';
 import { sendWhatsapp } from '@/services/sendWhatsapp';
-import { useCopilotLogic } from '@/services/useCopilotLogic';
+import { useCopilotLogic } from '@/hooks/useCopilotLogic';
 import { useNavigation, useFocusEffect } from '@react-navigation/native';
 import { useAuth } from '@/hooks/useAuth';
 import { fetchUserData } from '@/services/api/apiHomescreen';
