+++ react-native/hooks/useForgotPasswordLogic.tsx
@@ -1,25 +1,16 @@
-// useRegisterLogic.tsx
 import { useState, useEffect, useRef } from 'react';
 import { Alert } from 'react-native';
 import { useForm } from 'react-hook-form';
 import { useNavigation } from '@react-navigation/native';
-import { TextInput } from 'react-native-paper';
-import { ButtonSolid } from 'react-native-ui-buttons';
-import SupportButton from '@/components/SupportButton';
-import { createUser } from '@/services/api/apiRegister';
-import { validateCPF, checkPasswordMatch, checkEmptyFields, validateEmail } from '@/services/validateData';
+import { sendPasswordResetRequest } from '@/services/api/apiForgotPassword';
+import { checkEmptyFields, validateEmail } from '@/services/validateData';
 
 type FormData = {
   username: string;
-  password: string;
-  repeat_password: string;
-  fullname: string;
-  cpf: string;
-  telefone: string;
   email: string;
 };
 
-const useRegisterLogic = () => {
+const useForgotPasswordLogic = () => {
   const { register, setValue, handleSubmit: handleFormSubmit, setError, clearErrors, formState: { errors } } = useForm<FormData>();
   const navigation = useNavigation();
   const [isButtonDisabled, setIsButtonDisabled] = useState(false);
@@ -30,21 +21,17 @@ const useRegisterLogic = () => {
     let hasErrors = false;
 
     // Check for errors
-    let PasswordMatch = checkPasswordMatch(data.password, data.repeat_password, setError);
-    let ValidCPF = validateCPF(data.cpf, setError);
     let ValidEmail = validateEmail(data.email, setError);
     let NoEmptyFields = checkEmptyFields(data, setError);
     
-    hasErrors = PasswordMatch ||
-                ValidCPF ||
-                ValidEmail ||
+    hasErrors = ValidEmail ||
                 NoEmptyFields;
 
-    // If all validations pass, proceed with user creation
-    if (!hasErrors){
+    // If all validations pass, proceed with sending password reset request
+    if (!hasErrors) {
       try {
-        await createUser(data);
-        setIsButtonDisabled(false);
+        await sendPasswordResetRequest(data);
+        Alert.alert('E-mail enviado', 'Verifique sua caixa de spam ou de entrada para redefinir sua senha');
         navigation.goBack();
       } catch (error) {
         setIsButtonDisabled(false);
@@ -57,20 +44,10 @@ const useRegisterLogic = () => {
 
   useEffect(() => {
     register('username');
-    register('password');
-    register('repeat_password');
-    register('fullname');
-    register('cpf');
-    register('telefone');
     register('email');
   }, [register]);
 
   const usernameRef = useRef<TextInput>(null);
-  const passwordRef = useRef<TextInput>(null);
-  const repeatPasswordRef = useRef<TextInput>(null);
-  const fullnameRef = useRef<TextInput>(null);
-  const cpfRef = useRef<TextInput>(null);
-  const telefoneRef = useRef<TextInput>(null);
   const emailRef = useRef<TextInput>(null);
 
   return { 
@@ -78,15 +55,10 @@ const useRegisterLogic = () => {
     isButtonDisabled, 
     setValue, 
     usernameRef, 
-    passwordRef, 
-    repeatPasswordRef, 
-    fullnameRef, 
-    cpfRef, 
-    telefoneRef, 
     emailRef, 
     errors, 
     handleSubmit: handleFormSubmit 
   };
 };
 
-export default useRegisterLogic;
+export default useForgotPasswordLogic;
