+++ react-native/hooks/useRegisterLogic.tsx
@@ -1,17 +1,22 @@
+// useRegisterLogic.tsx
 import { useState, useEffect, useRef } from 'react';
 import { Alert } from 'react-native';
 import { useForm } from 'react-hook-form';
 import { useNavigation } from '@react-navigation/native';
-import { TextInput } from 'react-native-paper';
-import { sendPasswordResetRequest } from '@/services/api/apiForgotPassword';
+import { createUser } from '@/services/api/apiRegister';
 import { validateCPF, checkPasswordMatch, checkEmptyFields, validateEmail } from '@/services/validateData';
 
 type FormData = {
   username: string;
+  password: string;
+  repeat_password: string;
+  fullname: string;
+  cpf: string;
+  telefone: string;
   email: string;
 };
 
-const useForgotPasswordLogic = () => {
+const useRegisterLogic = () => {
   const { register, setValue, handleSubmit: handleFormSubmit, setError, clearErrors, formState: { errors } } = useForm<FormData>();
   const navigation = useNavigation();
   const [isButtonDisabled, setIsButtonDisabled] = useState(false);
@@ -22,17 +27,21 @@ const useForgotPasswordLogic = () => {
     let hasErrors = false;
 
     // Check for errors
+    let PasswordMatch = checkPasswordMatch(data.password, data.repeat_password, setError);
+    let ValidCPF = validateCPF(data.cpf, setError);
     let ValidEmail = validateEmail(data.email, setError);
     let NoEmptyFields = checkEmptyFields(data, setError);
     
-    hasErrors = ValidEmail ||
+    hasErrors = PasswordMatch ||
+                ValidCPF ||
+                ValidEmail ||
                 NoEmptyFields;
 
-    // If all validations pass, proceed with sending password reset request
-    if (!hasErrors) {
+    // If all validations pass, proceed with user creation
+    if (!hasErrors){
       try {
-        await sendPasswordResetRequest(data);
-        Alert.alert('E-mail enviado', 'Verifique sua caixa de spam ou de entrada para redefinir sua senha');
+        await createUser(data);
+        setIsButtonDisabled(false);
         navigation.goBack();
       } catch (error) {
         setIsButtonDisabled(false);
@@ -45,10 +54,20 @@ const useForgotPasswordLogic = () => {
 
   useEffect(() => {
     register('username');
+    register('password');
+    register('repeat_password');
+    register('fullname');
+    register('cpf');
+    register('telefone');
     register('email');
   }, [register]);
 
   const usernameRef = useRef<TextInput>(null);
+  const passwordRef = useRef<TextInput>(null);
+  const repeatPasswordRef = useRef<TextInput>(null);
+  const fullnameRef = useRef<TextInput>(null);
+  const cpfRef = useRef<TextInput>(null);
+  const telefoneRef = useRef<TextInput>(null);
   const emailRef = useRef<TextInput>(null);
 
   return { 
@@ -56,10 +75,15 @@ const useForgotPasswordLogic = () => {
     isButtonDisabled, 
     setValue, 
     usernameRef, 
+    passwordRef, 
+    repeatPasswordRef, 
+    fullnameRef, 
+    cpfRef, 
+    telefoneRef, 
     emailRef, 
     errors, 
     handleSubmit: handleFormSubmit 
   };
 };
 
-export default useForgotPasswordLogic;
+export default useRegisterLogic;
